Added logic to include session indicators in all management screens. (By moving the indicator from OFS/main.dtml to here.
Added support for the optional variable 'manage_tabs_message',
which can be used to include status messages. This somewhat experimental
feature is provided in hopes of doing away with message dialogs.

Changed the way that highlighting is done in tabs to account
for the fact that view actions may now be paths, rather than just
method names.

svn path=/Zope/trunk/; revision=16102