Changed INSTANCE_HOME to be = to SOFTWARE_HOME + 'var'.
Fixed a doc string that said S_H was the same as I_H

Added a Unix path to cheapen the experience and remove that nasty DOS
path.

If the log file doesn't exist, create instead of failing

Updates an INSTANCE_HOME + 'ZServer.pid' file with the process
id.  To convieniently kill ZServer.  yes this work with wondows.

svn path=/Zope/trunk/; revision=16296