Committing tempfile patch to branch.
This tempfile module is r1.27 from the Python MAIN CVS branch.  It fixes two rac
e conditions.

We will place this in lib/python and try to keep it up to date until we ditch Py
thon 1.52, at which time we can rely on the (patched) tempfile.py that ships wit
h the Python 2.X series.

svn path=/Zope/trunk/; revision=19231