Work around BTrees bug.
Some revision of the the BTrees code (possibly even the current head)
has a bug which causes (in this case) an OOBTree to lie about its
items.

Code like:

for key in oobtree.keys(): oobtree[key]

... will cause a KeyError when oobtree is asked to return the object
represented by key.

We work around this by explicitly detecting this problem and logging.

svn path=/Zope/trunk/; revision=22648