Get ZERO from the package.
Remove __version__

ZODB4 will have two new methods in the storage api:

    get_version() -> string
    set_version(string)

Prepare for this now by changing the packtime table into an info
table, which is just a key/value mapping for meta-information about
the storage.  For now, the only key we define is `packtime' although
we're reserving `version' for ZODB4.

svn path=/Zope/trunk/; revision=23411