Found and fixed the source of the "AttributeError: __error_log__" bug.
This code looks for the __error_log__ attribute in order to report an
error.  If it finds no such attribute, it reports the error in a
different way.  Unfortunately, the alternative branch was reporting
the exception that just happened to occur most recently, which is
irrelevant; it needs to report the error that was passed to it.
Fixed.

svn path=/Zope/trunk/; revision=24016