Warn when an attempt is made to have a security declaration on a nonexistent method. Removed one such method.
Fixed unclear security declarations. When bug 761 was fixed,
declareProtected(perm) was made illegal, at least one method name was
required. This checkin does the same for declarePrivate() and
declarePublic().

Also there was a bug in that a class having:
  security = ClassSecurityInfo()
  __ac_permissions__ = ((perm, ()),)
was not equivalent to a class having:
  __ac_permissions__ = ((perm, ()),)
This first form had as a buggy side effect have the behavior of
  __ac_permissions__ = ((perm, ('')),)
which is equivalent to
  __roles__ = PermissonRole(perm)
Not it will simply make the permission available from this object
(with default roles).


svn path=/Zope/branches/2.9/; revision=40279